Fixed admin login issue and added error handling for FAQ submission.
