+++ src/main/java/com/example/demo/FAQController.java
@@ -1,8 +1,10 @@
 package com.example.demo;
 
+import org.postgresql.util.PSQLException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -15,6 +17,36 @@ public class FAQController {
 
     @Autowired
     FAQService faqService;
+    private final String ADMIN_USERNAME = "admin";
+    private final String ADMIN_PASSWORD = "adminpassword";
+
+    @GetMapping("/admin-login")
+    public String showLoginPage() {
+        return "admin-login";
+    }
+
+    @PostMapping("/login")
+    public String processLogin(@RequestParam String username,
+                               @RequestParam String password, Model model) {
+        if (username.equals(ADMIN_USERNAME) && password.equals(ADMIN_PASSWORD)) {
+            return "redirect:/FAQ-database";
+        } else {
+            model.addAttribute("error", "Invalid login credentials");
+            return "redirect:/admin-login";
+        }
+    }
+
+    @ExceptionHandler(PSQLException.class)
+    public String psqlException(PSQLException psqlException) {
+        return "faq-submit-error";
+    }
+
+    @GetMapping("/FAQ-database")
+    public String showAllFAQs(Model model) {
+        List<FAQ> allFaqs = faqService.getAllFAQs();
+        model.addAttribute("allFaqs", allFaqs);
+        return "FAQ-database";
+    }
 
     @GetMapping("/faq")
     public String faq(@RequestParam(name = "category", defaultValue = "Practical") String category, Model model) {
