+++ src/main/resources/templates/admin-login.html
@@ -3,7 +3,6 @@
 <head>
     <meta charset="UTF-8">
     <title>Login</title>
-    <h1 class="propwise" th:style="font-size: 70px; text-decoration: underline; text-align: center; ">PropWise</h1>
     <style>
     body {
     background: #FFFAF0;
@@ -97,6 +96,7 @@
     </style>
 </head>
 <body>
+<h1 class="propwise">PropWise</h1>
 <div class="container">
 <!--    <h1 class="propwise">PropWise</h1>-->
     <h2>Admin Login to FAQ Database</h2>
