+++ src/main/resources/templates/faq-submit-error.html
@@ -1,0 +1,90 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Error Page</title>
+  <style>
+        body {
+            background-color: #FFFAF0;
+            color: #2F4F4F;
+            font-family: Arial, sans-serif;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            height: 100vh;
+            margin: 100px;
+            flex-direction: column;
+        }
+
+        .error-box {
+            border: 2px solid #2F4F4F;
+            padding: 20px;
+            border-radius: 10px;
+            box-shadow: 0 0 10px rgba(255, 255, 255, 0.5);
+            margin-bottom: 20px;
+        }
+
+        .error-box p {
+            color: #2F4F4F;
+        }
+
+        .button-container {
+      display: flex;
+      justify-content: space-between;
+      margin: 20px;
+    }
+    .button {
+      background-color: transparent;
+      color: #2F4F4F;
+      border: 1px solid #2F4F4F;
+      border-radius: 10px;
+      padding: 10px 20px;
+      cursor: pointer;
+      transition: background-color 0.3s, color 0.3s;
+    }
+    .button:hover {
+      background-color: #2F4F4F;
+      color: #FFFAF0 !important;
+    }
+        .help-button {
+      margin: 25px;
+      position: fixed;
+      top: 10px;
+      left: 100px;
+      background-color: transparent;
+      color: #2F4F4F;
+      border: 1px solid #2F4F4F;
+      border-radius: 10px;
+      padding: 10px 20px;
+      cursor: pointer;
+      transition: background-color 0.3s, color 0.3s;
+    }
+    .help-button:hover,
+     .button:hover a,
+      .help-button:hover a {
+      background-color: #2F4F4F;
+      color: #FFFAF0 !important;
+    }
+     .propwise {
+    font-size: 70px;
+    text-decoration: underline;
+    text-align: center;
+    }
+    </style>
+</head>
+<body>
+<h1 class="propwise">PropWise</h1>
+<div class="error-box">
+  <h1>Error!</h1>
+  <p>There was an error processing your question. It is likely that:</p>
+  <ol>
+    <li>You didn't input a full question. The form can only handle full questions. Go back and ensure you fill in a full question or exit the question submission page.</li>
+  </ol>
+</div>
+<div class="button-container">
+  <button class="button" onclick="window.location.href = 'form';"> &#8592; Back to Input</button>
+  <button class="button" onclick="window.location.href = 'launchpad';">🚀Launchpad</button>
+  <button class="help-button" onclick="window.location.href = 'help';">Help</button>
+</div>
+</body>
+</html>
